+++ src/Sql/Sql/ManagedDatabase Backup/Cmdlet/SetAzureSqlManagedDatabaseBackupLongTermRetentionPolicy.cs
@@ -16,6 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
+using Microsoft.Azure.Commands.Sql.Common;
 using Microsoft.Azure.Commands.Sql.Database.Model;
 using Microsoft.Azure.Commands.Sql.ManagedDatabaseBackup.Model;
 using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
@@ -113,6 +114,28 @@ namespace Microsoft.Azure.Commands.Sql.ManagedDatabaseBackup.Cmdlet
         public int WeekOfYear { get; set; }
 
         /// <summary>
+        /// Gets or sets whether time-based immutability is enabled for future LTR backups.
+        /// </summary>
+        [Parameter(Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "When set, future backups will have TimeBasedImmutability enabled.")]
+        [ValidateSet(
+            DatabaseBackupConstants.TimeBasedImmutabilityValues.Enabled,
+            DatabaseBackupConstants.TimeBasedImmutabilityValues.Disabled)]
+        public string TimeBasedImmutability { get; set; }
+
+        /// <summary>
+        /// Gets or sets the time-based immutability mode for future LTR backups.
+        /// </summary>
+        [Parameter(Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The setting for time-based immutability mode for future backups. Only effective if TimeBasedImmutability is enabled. Value can be either Locked or Unlocked. Caution: Immutability of LTR backup cannot be removed if TimeBasedImmutabilityMode is Locked.")]
+        [ValidateSet(
+            DatabaseBackupConstants.TimeBasedImmutabilityModeValues.Unlocked,
+            DatabaseBackupConstants.TimeBasedImmutabilityModeValues.Locked)]
+        public string TimeBasedImmutabilityMode { get; set; }
+
+        /// <summary>
         /// Get the entities from the service
         /// </summary>
         /// <returns>The list of entities</returns>
@@ -159,6 +182,12 @@ namespace Microsoft.Azure.Commands.Sql.ManagedDatabaseBackup.Cmdlet
                 WeekOfYear = 1;
             }
 
+            if (!string.IsNullOrEmpty(TimeBasedImmutabilityMode) &&
+                (string.IsNullOrEmpty(TimeBasedImmutability) || TimeBasedImmutability == DatabaseBackupConstants.TimeBasedImmutabilityValues.Disabled))
+            {
+                throw new ArgumentException("TimeBasedImmutabilityMode can only be specified if TimeBasedImmutability is set to Enabled");
+            }
+
             return new List<AzureSqlManagedDatabaseBackupLongTermRetentionPolicyModel>()
             {
                 new AzureSqlManagedDatabaseBackupLongTermRetentionPolicyModel()
@@ -170,7 +199,9 @@ namespace Microsoft.Azure.Commands.Sql.ManagedDatabaseBackup.Cmdlet
                     WeeklyRetention = WeeklyRetention,
                     MonthlyRetention = MonthlyRetention,
                     YearlyRetention = YearlyRetention,
-                    WeekOfYear = WeekOfYear
+                    WeekOfYear = WeekOfYear,
+                    TimeBasedImmutability = TimeBasedImmutability,
+                    TimeBasedImmutabilityMode = TimeBasedImmutabilityMode
                 }
             };
         }