+++ src/Sql/Sql/ManagedDatabase Backup/Services/AzureSqlManagedDatabaseBackupAdapter.cs
@@ -230,7 +230,9 @@ namespace Microsoft.Azure.Commands.Sql.ManagedDatabaseBackup.Services
                 WeeklyRetention = response.WeeklyRetention,
                 MonthlyRetention = response.MonthlyRetention,
                 YearlyRetention = response.YearlyRetention,
-                WeekOfYear = response.WeekOfYear
+                WeekOfYear = response.WeekOfYear,
+                TimeBasedImmutability = response.TimeBasedImmutability,
+                TimeBasedImmutabilityMode = response.TimeBasedImmutabilityMode
             };
         }
 
@@ -257,7 +259,9 @@ namespace Microsoft.Azure.Commands.Sql.ManagedDatabaseBackup.Services
                         WeeklyRetention = model.WeeklyRetention,
                         MonthlyRetention = model.MonthlyRetention,
                         YearlyRetention = model.YearlyRetention,
-                        WeekOfYear = model.WeekOfYear
+                        WeekOfYear = model.WeekOfYear,
+                        TimeBasedImmutability = model.TimeBasedImmutability,
+                        TimeBasedImmutabilityMode = model.TimeBasedImmutabilityMode
                     });
             return new AzureSqlManagedDatabaseBackupLongTermRetentionPolicyModel()
             {
@@ -267,7 +271,9 @@ namespace Microsoft.Azure.Commands.Sql.ManagedDatabaseBackup.Services
                 WeeklyRetention = response.WeeklyRetention,
                 MonthlyRetention = response.MonthlyRetention,
                 YearlyRetention = response.YearlyRetention,
-                WeekOfYear = response.WeekOfYear
+                WeekOfYear = response.WeekOfYear,
+                TimeBasedImmutability = response.TimeBasedImmutability,
+                TimeBasedImmutabilityMode = response.TimeBasedImmutabilityMode
             };
         }
 
@@ -280,6 +286,10 @@ namespace Microsoft.Azure.Commands.Sql.ManagedDatabaseBackup.Services
                 BackupExpirationTime = backup.BackupExpirationTime,
                 BackupName = backup.Name,
                 BackupStorageRedundancy = backup.BackupStorageRedundancy,
+                IsBackupImmutable = backup.IsBackupImmutable,
+                TimeBasedImmutability = backup.TimeBasedImmutability,
+                TimeBasedImmutabilityMode = backup.TimeBasedImmutabilityMode,
+                LegalHoldImmutability = backup.LegalHoldImmutability,
                 BackupTime = backup.BackupTime,
                 DatabaseDeletionTime = backup.DatabaseDeletionTime,
                 DatabaseName = backup.DatabaseName,